hotfix: 로직 수정
