--- src/supabase/.gitignore
@@ -1,5 +1,0 @@
-# Supabase
-.branches
-.temp
-.env
-.env.local
