--- src/supabase/functions/_shared/cors.ts
@@ -1,4 +1,0 @@
-export const corsHeaders = {
-  'Access-Control-Allow-Origin': '*',
-  'Access-Control-Allow-Headers': 'authorization, x-client-info, apikey, content-type',
-}
