+++ supabase/functions/payment/index.ts
@@ -1,27 +1,27 @@
-import "jsr:@supabase/functions-js/edge-runtime.d.ts"
-import { corsHeaders } from '../_shared/cors.ts'
+import "jsr:@supabase/functions-js/edge-runtime.d.ts";
+import { corsHeaders } from '../_shared/cors.ts';
 
-const widgetSecretKey = Deno.env.get('WIDGET_SECRET_KEY')
+const widgetSecretKey = Deno.env.get('WIDGET_SECRET_KEY');
 
 Deno.serve(async (req) => {
-  if (req.method === "POST" && req.url === "http://edge-runtime.supabase.com/payment") {
+  console.log(req.method);
+  if (req.method === 'OPTIONS') {
+    return new Response('ok', { headers: corsHeaders })
+  }
+  // Handle the actual POST request
+  if (req.method === "POST" && req.url.includes("/payment")) {
     const { paymentKey, orderId, amount } = await req.json();
+    const encryptedSecretKey = "Basic " + btoa(widgetSecretKey + ":");
 
-    const encryptedSecretKey =
-      "Basic " + btoa(widgetSecretKey + ":");
-      try {
-        const response = await fetch("https://api.tosspayments.com/v1/payments/confirm", {
-          method: "POST", // 메서드 추가
-          headers: {
-            Authorization: encryptedSecretKey, // 헤더에 encryptedSecretKey 추가
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify({ // body에 JSON 문자열로 전달
-            orderId: orderId,
-            amount: amount,
-            paymentKey: paymentKey,
-          }),
-        });
+    try {
+      const response = await fetch("https://api.tosspayments.com/v1/payments/confirm", {
+        method: "POST",
+        headers: {
+          Authorization: encryptedSecretKey,
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({ orderId, amount, paymentKey }),
+      });
 
         // data.json()으로 받아서 비동기 처리를 해줘야 한다.
         const data = await response.json();
